refactor: consolidate PK/FK domain constraints into constraints.py
