+++ docs/todo/todo.md
@@ -6,7 +6,7 @@
 - [ ] make partitioned_by a Column-level thing on api DeltaTable
 - [ ] add unique columns: ALTER TABLE U ADD CONSTRAINT u_uq_email UNIQUE(email);
 - [ ] should all of the sql statements live in a dedicated file and be imported into reader/exeecutor?
-- [ ] Make the application layer backend-agnostic, not just backend-library-free. Today the `backend-imports-stay-in-adapters` contract only forbids importing the `delta`/`pyspark` *packages*; the application layer still hardcodes Delta-family *policy* — the safe-change rules (`PartitioningChangeNotSupported`, `ColumnDataTypeChangeNotSupported`, `NonNullableColumnAdd`, `ColumnMappingRequiredForDrop`) exist because Databricks/Delta rejects those operations, and `application/properties.py` holds the Delta property catalogue (`DELTA_PROPERTY_REGISTRY`, `COLUMN_MAPPING_MODE_KEY`, `delta.*` key strings) consumed by `validation.py`, `api/properties.py`, and the reader. The layering is sound for now (application is the only layer all three consumers can reach without violating the `adapters|api → application → domain` order), but it means "backend-agnostic" is only half true. The real refactor: `application` holds only the *mechanism* (how rules run, the `Rule` protocol, how `validate_diff` composes them), and the backend *specifics* (which rules, which property keys and their restrictions) are injected from the composition root or an adapter-provided port. This must move ALL the backend-flavoured rules and the property catalogue together — moving only the property pieces just relocates a third of the problem. Larger change, own PR. See the property-ownership spec addendum for how the registry ended up in application.
+- [ ] Make the application layer backend-agnostic, not just backend-library-free. Today the `backend-imports-stay-in-adapters` contract only forbids importing the `delta`/`pyspark` _packages_; the application layer still hardcodes Delta-family _policy_ — the safe-change rules (`PartitioningChangeNotSupported`, `ColumnDataTypeChangeNotSupported`, `NonNullableColumnAdd`, `ColumnMappingRequiredForDrop`) exist because Databricks/Delta rejects those operations, and `application/properties.py` holds the Delta property catalogue (`DELTA_PROPERTY_REGISTRY`, `COLUMN_MAPPING_MODE_KEY`, `delta.*` key strings) consumed by `validation.py`, `api/properties.py`, and the reader. The layering is sound for now (application is the only layer all three consumers can reach without violating the `adapters|api → application → domain` order), but it means "backend-agnostic" is only half true. The real refactor: `application` holds only the _mechanism_ (how rules run, the `Rule` protocol, how `validate_diff` composes them), and the backend _specifics_ (which rules, which property keys and their restrictions) are injected from the composition root or an adapter-provided port. This must move ALL the backend-flavoured rules and the property catalogue together — moving only the property pieces just relocates a third of the problem. Larger change, own PR. See the property-ownership spec addendum for how the registry ended up in application.
 - [ ] Think of whether to make DeltaTable automatically make columns unique if they are PK cols
 - [ ] make the nested if else statments in engine more readable
 - [ ] do we need backticked things in sql compiler or can we do without?
@@ -14,17 +14,17 @@
 - [ ] consider replacing typle comps in validation with regular loops for readability
 - [ ] Test hygiene sweep (low-risk, deferred from FK PR review): (1) rename ~60 abbreviated test locals — `reg`→registry, `fq`→fully_qualified_name, the `qn` fixture in test_reader.py→qualified_name, `tr`→table_run, `vf`/`sr`/`col`/`t`/`d`, module consts `_QN`/`_COL`; (2) fill in ~25 bare wordless Given/When/Then markers; (3) rename 4 meta-label/stale-class test names (`_happy_path` suffixes, `test_table_snapshot_...` that actually tests DesiredTable); (4) add GWT structure to ~12 tests in test_preview/test_dialect/test_types that lack it. (Themes 3/4/5 — private-method coupling, IntEnum-value asserts, implementation-string asserts — already fixed in the FK PR.)
 - [ ] Review AnalysisException catches in executor and reader
-- [ ] Review foreign key logic
+- [x] Review foreign key logic
 - [ ] Simplify `_fetch_foreign_keys` in the reader: replace the stringly-typed `dict[str, dict]` grouping with `itertools.groupby` (the query already does `ORDER BY constraint_name, ordinal_position`, so rows are contiguous) plus a named `_foreign_key_from_rows(constraint_name, rows)` helper that reads local/referenced columns in ordinal order and takes the referenced table from the first row
-- [ ] utilise __init__ __all__ so that we can reduce the number of import line
+- [ ] utilise **init** **all** so that we can reduce the number of import line
 - [ ] Add build docs to pre-commit?
 - [ ] should report.any_failures be report.has_any_failures?
 - [ ] should DesiredTableSource live in ports.py?
 - [ ] Decide whether to emit `RELY` on FK/PK constraints. Without `RELY`, Databricks treats informational constraints as documentation only and the optimizer cannot use them for join elimination / query rewrite. If the point of declaring keys is optimization, they are currently inert. Weigh against the risk of `RELY` on unverified data (the optimizer trusts it). Would add a `rely: bool` to `ForeignKeyConstraint`/`PrimaryKeyConstraint` and a ` NOT ENFORCED RELY` / ` NOT ENFORCED` suffix in the compiler.
 - [ ] Add a test for the cycle + invalid-FK-target combination in resolve(): a table that is both on a foreign key cycle AND references a non-primary-key column currently reports REFERENCED_COLUMNS_NOT_A_KEY (the FK->PK check runs before the CYCLE check, by design). No test covers this interaction; document the precedence or add coverage.
-- [ ] investigate TableSnapshot.__post_init__(self) on DesiredTable. Should all validation be in DesiredTable?
+- [ ] investigate TableSnapshot.**post_init**(self) on DesiredTable. Should all validation be in DesiredTable?
 - [ ] look for redundancies; things like frozenset for local variables
-- [ ] review if we need from __future__ import annotations
+- [ ] review if we need from **future** import annotations
 - [ ] review types hints; things like Mapping, AbstractSet, etc
 - [ ] review Matched and Matched.common, particularly in _diff_columns
 - [ ] Add tags, comment, schema, etc as properties to DeltaTable
