+++ src/delta_engine/application/dependency_resolution.py
@@ -39,7 +39,7 @@ from dataclasses import dataclass
 
 from delta_engine.application.failures import ForeignKeyFailure, ForeignKeyFailureReason
 from delta_engine.domain.model import QualifiedName
-from delta_engine.domain.model.foreign_key import ForeignKeyConstraint
+from delta_engine.domain.model.constraints import ForeignKeyConstraint
 from delta_engine.domain.model.table import DesiredTable
 
 # A table dependency graph is small (tens of tables, shallow chains), so the
