+++ src/delta_engine/domain/model/constraints.py
@@ -1,4 +1,4 @@
-"""Domain value object representing a foreign key constraint."""
+"""Domain value objects representing key constraints (primary and foreign)."""
 
 from __future__ import annotations
 
@@ -6,6 +6,42 @@ from dataclasses import dataclass
 from typing import Self
 
 from delta_engine.domain.model.qualified_name import QualifiedName
+
+
+@dataclass(frozen=True, slots=True)
+class PrimaryKeyConstraint:
+    """
+    A primary key constraint declaration.
+
+    Symmetric with :class:`ForeignKeyConstraint`: a table-level key constraint
+    over an ordered set of columns.
+
+    Attributes:
+        columns: Ordered tuple of column names that make up the primary key.
+        constraint_name: The constraint name. Generated by the API layer when a
+            ``DeltaTable`` is lowered to a ``DesiredTable`` (``{table}_pk``), and
+            read from the catalog for observed constraints.
+
+    """
+
+    columns: tuple[str, ...]
+    constraint_name: str
+
+    def __post_init__(self) -> None:
+        if not self.columns:
+            raise ValueError("columns must not be empty")
+        seen: set[str] = set()
+        for column in self.columns:
+            if column in seen:
+                raise ValueError(f"Duplicate primary key column: {column}")
+            seen.add(column)
+        if not self.constraint_name.strip():
+            raise ValueError("constraint_name must not be blank")
+
+    @classmethod
+    def generate(cls, *, table_name: str, columns: tuple[str, ...]) -> Self:
+        """Return a constraint over ``columns`` named ``{table_name}_pk``."""
+        return cls(columns=columns, constraint_name=f"{table_name}_pk")
 
 
 @dataclass(frozen=True, slots=True)
