--- src/delta_engine/domain/model/primary_key.py
@@ -1,42 +1,0 @@
-"""Domain value object representing a primary key constraint."""
-
-from __future__ import annotations
-
-from dataclasses import dataclass
-from typing import Self
-
-
-@dataclass(frozen=True, slots=True)
-class PrimaryKeyConstraint:
-    """
-    A primary key constraint declaration.
-
-    Symmetric with :class:`ForeignKeyConstraint`: a table-level key constraint
-    over an ordered set of columns.
-
-    Attributes:
-        columns: Ordered tuple of column names that make up the primary key.
-        constraint_name: The constraint name. Generated by the API layer when a
-            ``DeltaTable`` is lowered to a ``DesiredTable`` (``{table}_pk``), and
-            read from the catalog for observed constraints.
-
-    """
-
-    columns: tuple[str, ...]
-    constraint_name: str
-
-    def __post_init__(self) -> None:
-        if not self.columns:
-            raise ValueError("columns must not be empty")
-        seen: set[str] = set()
-        for column in self.columns:
-            if column in seen:
-                raise ValueError(f"Duplicate primary key column: {column}")
-            seen.add(column)
-        if not self.constraint_name.strip():
-            raise ValueError("constraint_name must not be blank")
-
-    @classmethod
-    def generate(cls, *, table_name: str, columns: tuple[str, ...]) -> Self:
-        """Return a constraint over ``columns`` named ``{table_name}_pk``."""
-        return cls(columns=columns, constraint_name=f"{table_name}_pk")
