+++ tests/domain/plan/test_diff.py
@@ -11,8 +11,7 @@ from delta_engine.domain.model import (
     String,
     TableAspect,
 )
-from delta_engine.domain.model.foreign_key import ForeignKeyConstraint
-from delta_engine.domain.model.primary_key import PrimaryKeyConstraint
+from delta_engine.domain.model.constraints import ForeignKeyConstraint, PrimaryKeyConstraint
 from delta_engine.domain.plan.actions import (
     ActionPlan,
     AddColumn,
